refactor: Moved tracker files to appropriate locations
