+++ lib/src/tracker/mod.rs
@@ -1,16 +1,15 @@
 use anyhow::Result;
 use serde::{
-   Deserialize, Serialize,
+   Deserialize,
    de::{self, Visitor},
 };
 use std::{fmt, net::Ipv4Addr};
+mod http;
+mod udp;
+// mod websocket;
 
-#[derive(Debug, Deserialize)]
-pub struct TrackerResponse {
-   pub interval: usize,
-   #[serde(deserialize_with = "deserialize_peers")]
-   pub peers: Vec<Peer>,
-}
+use http::*;
+// use udp::*;
 
 #[derive(Debug, Deserialize)]
 pub struct Peer {
@@ -23,39 +22,13 @@ pub struct Peer {
 /// Example: udp://tracker.opentrackr.org:1337/announce
 #[derive(Debug)]
 pub enum AnnounceUri {
+   /// HTTP Spec
+   /// https://www.bittorrent.org/beps/bep_0003.html
    Http(String),
+   /// UDP Spec
+   /// https://www.bittorrent.org/beps/bep_0015.html
    Udp(String),
    Websocket(String),
-}
-
-fn deserialize_peers<'de, D>(deserializer: D) -> Result<Vec<Peer>, D::Error>
-where
-   D: serde::Deserializer<'de>,
-{
-   let bytes = Vec::<u8>::deserialize(deserializer).expect("Invalid bytes");
-
-   let mut peers = Vec::new();
-   for chunk in bytes.chunks(6) {
-      if chunk.len() != 6 {
-         return Err(de::Error::custom("Invalid peer chunk length"));
-      }
-      let ip = Ipv4Addr::new(chunk[0], chunk[1], chunk[2], chunk[3]);
-
-      let port = u16::from_be_bytes([chunk[4], chunk[5]]);
-      peers.push(Peer { ip, port });
-   }
-
-   Ok(peers)
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-struct TrackerRequest {
-   peer_id: String,
-   port: u16,
-   uploaded: u8,
-   downloaded: u8,
-   left: u8,
-   compact: u8,
 }
 
 impl AnnounceUri {
