+++ lib/src/tracker/udp.rs
@@ -1,0 +1,8 @@
+/// 2 bytes for source port, 2 bytes for destination port, 2 bytes for length, and 2 bytes for checksum
+/// https://en.wikipedia.org/wiki/User_Datagram_Protocol
+/// https://www.bittorrent.org/beps/bep_0015.html
+use anyhow::Result;
+
+pub async fn get() -> Result<()> {
+   Ok(())
+}
